+++ src/main/java/com/example/library_IAU/repository/UsersRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<UsersModel, Integer> {
+public interface UsersRepository extends JpaRepository<UsersModel, Long> {
 
     Optional<UsersModel> findByEmailAndPassword(String email, String password);
+
 }
